move Image component to own folder
