+++ src/components/Images/Images.js
@@ -1,6 +1,6 @@
 import React, { PropTypes as T } from 'react';
 
-import Image from './Image';
+import Image from '../Image/Image';
 
 export default function Images({
   photos,
